// packages/app/src/views/pricing/types.ts
export enum PaymentTypeEnum {
  BASIC_PLAN_MONTHLY = 'BASIC_PLAN_MONTHLY',
  BASIC_PLAN_YEARLY = 'BASIC_PLAN_YEARLY',
  STANDARD_PLAN_MONTHLY = 'STANDARD_PLAN_MONTHLY',
  STANDARD_PLAN_YEARLY = 'STANDARD_PLAN_YEARLY',
  PROFESSIONAL_PLAN_MONTHLY = 'PROFESSIONAL_PLAN_MONTHLY',
  PROFESSIONAL_PLAN_YEARLY = 'PROFESSIONAL_PLAN_YEARLY',
  ENTERPRISE_PLAN_MONTHLY = 'ENTERPRISE_PLAN_MONTHLY',
  ENTERPRISE_PLAN_YEARLY = 'ENTERPRISE_PLAN_YEARLY',
  EXTRA_PLAN_LOW = 'EXTRA_PLAN_LOW',
  EXTRA_PLAN_MIDDLE = 'EXTRA_PLAN_MIDDLE',
  EXTRA_PLAN_HIGH = 'EXTRA_PLAN_HIGH',
}

export enum RechargeTypeEnum {
  /** 月付 */
  MONTHLY = 'monthly',
  /** 年付 */
  YEARLY = 'yearly',
  /** unknown */
  REPLENISH = 'replenish',
}